refactor: data와 domain model의 comment 패키지 분리
