+++ android/app/src/main/java/com/zzang/chongdae/data/remote/mapper/comment/CommentCreatedAtMapper.kt
@@ -1,7 +1,9 @@
-package com.zzang.chongdae.data.remote.mapper
+package com.zzang.chongdae.data.remote.mapper.comment
 
 import com.zzang.chongdae.data.remote.dto.response.comment.CommentCreatedAtResponse
-import com.zzang.chongdae.domain.model.CommentCreatedAt
+import com.zzang.chongdae.data.remote.mapper.toLocalDate
+import com.zzang.chongdae.data.remote.mapper.toLocalTime
+import com.zzang.chongdae.domain.model.comment.CommentCreatedAt
 
 fun CommentCreatedAtResponse.toDomain(): CommentCreatedAt {
     return CommentCreatedAt(
