+++ android/app/src/main/java/com/zzang/chongdae/data/remote/mapper/comment/CommentOfferingInfoMapper.kt
@@ -1,7 +1,7 @@
-package com.zzang.chongdae.data.remote.mapper
+package com.zzang.chongdae.data.remote.mapper.comment
 
 import com.zzang.chongdae.data.remote.dto.response.comment.CommentOfferingInfoResponse
-import com.zzang.chongdae.domain.model.CommentOfferingInfo
+import com.zzang.chongdae.domain.model.comment.CommentOfferingInfo
 
 fun CommentOfferingInfoResponse.toDomain() =
     CommentOfferingInfo(
