+++ android/app/src/main/java/com/zzang/chongdae/data/repository/CommentDetailRepositoryImpl.kt
@@ -3,11 +3,11 @@ package com.zzang.chongdae.data.repository
 import com.zzang.chongdae.common.handler.DataError
 import com.zzang.chongdae.common.handler.Result
 import com.zzang.chongdae.data.remote.dto.request.CommentRequest
-import com.zzang.chongdae.data.remote.mapper.toDomain
+import com.zzang.chongdae.data.remote.mapper.comment.toDomain
 import com.zzang.chongdae.data.source.comment.CommentRemoteDataSource
 import com.zzang.chongdae.di.annotations.CommentDetailDataSourceQualifier
-import com.zzang.chongdae.domain.model.Comment
-import com.zzang.chongdae.domain.model.CommentOfferingInfo
+import com.zzang.chongdae.domain.model.comment.Comment
+import com.zzang.chongdae.domain.model.comment.CommentOfferingInfo
 import com.zzang.chongdae.domain.repository.CommentDetailRepository
 import javax.inject.Inject
