--- android/app/src/main/java/com/zzang/chongdae/domain/model/Participation.kt
@@ -1,6 +1,0 @@
-package com.zzang.chongdae.domain.model
-
-data class Participation(
-    val offeringCondition: OfferingCondition,
-    val currentCount: CurrentCount,
-)
