+++ android/app/src/main/java/com/zzang/chongdae/domain/model/comment/Comment.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.domain.model
+package com.zzang.chongdae.domain.model.comment
 
 data class Comment(
     val content: String,
