+++ android/app/src/main/java/com/zzang/chongdae/domain/model/comment/CommentCreatedAt.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.domain.model
+package com.zzang.chongdae.domain.model.comment
 
 import java.time.LocalDate
 import java.time.LocalTime
