+++ android/app/src/main/java/com/zzang/chongdae/domain/model/comment/CommentOfferingInfo.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.domain.model
+package com.zzang.chongdae.domain.model.comment
 
 class CommentOfferingInfo(
     val status: String,
