+++ android/app/src/main/java/com/zzang/chongdae/domain/model/participant/Participation.kt
@@ -1,0 +1,9 @@
+package com.zzang.chongdae.domain.model.participant
+
+import com.zzang.chongdae.domain.model.CurrentCount
+import com.zzang.chongdae.domain.model.OfferingCondition
+
+data class Participation(
+    val offeringCondition: OfferingCondition,
+    val currentCount: CurrentCount,
+)
