+++ android/app/src/main/java/com/zzang/chongdae/domain/repository/CommentDetailRepository.kt
@@ -2,8 +2,8 @@ package com.zzang.chongdae.domain.repository
 
 import com.zzang.chongdae.common.handler.DataError
 import com.zzang.chongdae.common.handler.Result
-import com.zzang.chongdae.domain.model.Comment
-import com.zzang.chongdae.domain.model.CommentOfferingInfo
+import com.zzang.chongdae.domain.model.comment.Comment
+import com.zzang.chongdae.domain.model.comment.CommentOfferingInfo
 
 interface CommentDetailRepository {
     suspend fun saveComment(
