+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/model/comment/CommentUiModel.kt
@@ -1,6 +1,6 @@
 package com.zzang.chongdae.presentation.view.commentdetail.model.comment
 
-import com.zzang.chongdae.domain.model.Comment
+import com.zzang.chongdae.domain.model.comment.Comment
 import com.zzang.chongdae.presentation.util.toFormattedDate
 import com.zzang.chongdae.presentation.util.toFormattedTime
 import com.zzang.chongdae.presentation.view.commentdetail.adapter.comment.CommentViewType
